Don't use the UINT8 type with va_arg.
Rev: src/format.c:1.18

+++ src/format.c
@@ -102,7 +102,7 @@ UINT32 ssh_vformat_length(char *f, va_list args)
 	      break;
 
 	    case 'c':
-	      (void) va_arg(args, UINT8);
+	      (void) va_arg(args, int);
 	      /* Fall through */
 	    case '%':
 	      f++;
